Create Ethical-BusinessPlan_Images
FinBuddy: Our mission is to enable young professionals and students to reach their long-term objectives and make wise financial decisions. We came up with the idea because we saw a gap in this demographic's access to individualised financial counsel. Our goal is to offer a complete platform that makes managing finances easier, teaches people how to be frugal, and promotes financial independence.